added Pytorch and other module
