+++ templates/index.html
@@ -7,8 +7,12 @@
 </head>
 <body>
     <form action = "/input" method = "POST">
-        <p>Enter Genre:</p>
-        <p><input type = "text" name = "genre"></p>
+        <p>Choose Starting Scene:
+                Dune
+                Star Wars
+                Hitchhikers
+        </p>
+        <p><input type = "text" name = "choice"></p>
         <p><input type = "submit" value = "submit"></p>
     </form>
