+++ templates/results.html
@@ -3,12 +3,11 @@
 <head>
     <meta charset="UTF-8">
     <title>Pick-It Movies</title>
-
 </head>
 <body>
    <p>"these are your {{results3}}</p>
-   <p></p>
-   <p></p>
-
+    <!-- {% for item in results3 %}
+    <p>"{{item}}</p>
+    {% endfor %} -->
 </body>
 </html>
